add update speed to refresh command
Writing a valid update speed value to register address 0x01 will refresh
the ePaper display using the strings stored in memory. The value 0x00 is
a full refresh with no ghosting, 0x03 is the fasted partical refresh
which will exhibit ghosting. Values in between the two are a mix of
speed versus fidelity.

Signed-off-by: Mike Szczys <mike@golioth.io>