add slideshow modules
Slides show a label and a value on the screen. Slides are added at
runtime by assigning a unique id and a text label. Once added the value
may be updated as needed using the id.

Pass a delay in milliseconds to the start(<delay>) function. It will
immediately update the display, then use a periodic timer to step
through the available slides.

Signed-off-by: Mike Szczys <mike@golioth.io>