override splashscreen by loading a bin
The Golioth logo is displayed as the default splashscreen. This commit
adds the option to use Thonny to store a splashscreen.bin file on
Ostentus. If this file is present it will be used whenever show_splash()
is called.

Images but be 5000 byte binary files. Create a black and white image
that is 200x200 pixels, then invert the colors. Use convert.py found in
the pimoroni-pico repo to conver the image:
https://github.com/szczys/pimoroni-pico/tree/main/examples/badger2040/image_converter

```
python3 convert.py --binary source_image.png
```

Signed-off-by: Mike Szczys <mike@golioth.io>