service slideshow from ostentus loop
Monitoring the i2c buffer is the highest priority task. This commit
moves the slideshow timing out of an interrupt-based execution so the
screen refresh doesn't block the i2c buffer monitor. The timer now sets
a flag that is checked from the ostentus loop when there are no i2c
messgaes pending.

Signed-off-by: Mike Szczys <mike@golioth.io>