service cap touch in loop with LED indicator
When cap touch is used during slideshows, the input will set a flag that
is then serviced fro the ostentus loop. This introduces a lot of latency
if the screen is being frequently updated (full or partial) so the User
LED will be illuminated when the press is detected and shut off after
the update has occurred.

Signed-off-by: Mike Szczys <mike@golioth.io>