handle i2c when Ostentus is missing
The Zephyr lib will check to see if Ostentus is present by clearing the
memory the first time an i2c operation is requested. If Ostentus is not
found, an error message will be raised. All future i2c write requests to
Ostentus will result in a debug message (the i2c write operation will
not be attempted to avoid repeated error messages from the i2c driver).

Signed-off-by: Mike Szczys <mike@golioth.io>