abstract partial_update into data and execute
The update command for partial_updates applies to the entire screen.
This commit separates that command from the data write step so that
several smaller parts of the screen may be updated in the ePaper memory
before issuing the partial update command.

This change has not been connected to the Python layer yet.

Signed-off-by: Mike Szczys <mike@golioth.io>