add timer to delay summary parital updates
When in summary mode, the partial update was being triggered for every
new data value on the first three pages. So if three readings (eg:
temperature, pressure, humitidy) all come in at the same time, the
screeen was doing a partial update three times. Because each partial
update degrades the existing image fidelity, the results looked poor.

This commit uses a 250 millisecond timer to delay partial updates that
are triggered by incoming data. This way, several pieces of data may be
processed, but only one partial update is performed.

Signed-off-by: Mike Szczys <mike@golioth.io>