add hardware reset to partial update
flickering and screen degredation on partial are due to a missing
hardware reset before beginning the memory writes. Technically we only
need to reset and border freeze once for the first partial write (before
issuing the partial update command) but it works with these on every
partial memroy write, which also makes the code a bit simpler.

Signed-off-by: Mike Szczys <mike@golioth.io>