move buffer full check for after immediate process
Wait to check for a full buffer until after data is received and process
immediate registers have been serviced.

Since the process immediate register addresses are never put into the
fifo, they can still be received even when the fifo is full. This is
necessary for an upcoming enhancement that implements resetting the
Ostentus. Without this change, reset wouldn't be possible if the fifo is
already full.

Signed-off-by: Mike Szczys <mike@golioth.io>